requirements.txt
cffi==1.11.2
pycparser==2.18
pygit2==0.26.0
PyYAML==3.11
requests==2.9.1
six==1.11.0
virtualenv==15.1.0
